basiciris.txt
This iris data file except with one of the species of flower removed.  I also mixed things up so that a training set working from an initial segment will have samples of both species.  A training set of 10 achieves perfect accuracy, but more or less results in errors.  It seems to be all or nothing.  Put this in your desktop or matlab folder to try it out.